新增：加了输入条件判定，增强程序健壮性

+++ service/groupService.js
@@ -23,23 +23,37 @@ exports.createGroup = async function(userId,groupName){
     return Promise.resolve(jsonUtils.getResponseBody(codes.groupname_exists_error))
 }
 exports.setGroupNum = async function(userId,friendId,groupId){
-    let value = null
-    try {
-        value = await groupRepository.changeGroupNum(userId,friendId,groupId)
-    } catch (e) {
-        return Promise.reject(jsonUtils.getResponseBody(codes.other_error, e))
+    let is_friend = null
+    is_friend = await groupRepository.isFriend(userId,friendId)
+    if(is_friend.length===1) {
+        let is_group_existed = null
+        is_group_existed=await groupRepository.isGroupExisted(groupId)
+        if(is_group_existed.length===1) {
+            let value = null
+            try {
+                value = await groupRepository.changeGroupNum(userId, friendId, groupId)
+            } catch (e) {
+                return Promise.reject(jsonUtils.getResponseBody(codes.other_error, e))
+            }
+            return Promise.resolve(jsonUtils.getResponseBody(codes.success))
+        }
+        return Promise.resolve(jsonUtils.getResponseBody(codes.group_not_existed))
     }
-
-    return Promise.resolve(jsonUtils.getResponseBody(codes.success))
+    return Promise.resolve(jsonUtils.getResponseBody(codes.not_friend_build_group_error))
 }
 exports.deleteGroup = async function(groupId){
-    let value = null
-    try {
-        value = await groupRepository.deleteGroup(groupId)
-    } catch (e) {
-        return Promise.reject(jsonUtils.getResponseBody(codes.other_error, e))
+    let is_group_existed = null
+    is_group_existed=await groupRepository.isGroupExisted(groupId)
+    if(is_group_existed.length===1) {
+        let value = null
+        try {
+            value = await groupRepository.deleteGroup(groupId)
+        } catch (e) {
+            return Promise.reject(jsonUtils.getResponseBody(codes.other_error, e))
+        }
+        return Promise.resolve(jsonUtils.getResponseBody(codes.success))
     }
-    return Promise.resolve(jsonUtils.getResponseBody(codes.success))
+    return Promise.resolve(jsonUtils.getResponseBody(codes.group_not_existed))
 }
 exports.findAllGroup = async function(userId){
     let value = null
@@ -48,7 +62,7 @@ exports.findAllGroup = async function(userId){
     } catch (e) {
         return Promise.reject(jsonUtils.getResponseBody(codes.other_error, e))
     }
-    if (value.length === 0) { //获取到的用户数量为0：用户不存在
+    if (value.length === 0) { //获取到小组数为0
         return Promise.reject(jsonUtils.getResponseBody(codes.other_error, null))
     }
     return Promise.resolve(jsonUtils.getResponseBody(codes.success))
